requirements.txt
streamlit==1.31.0
pandas==2.1.4
redis==5.0.1
resend
celery==5.3.6
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.114.0
seaborn==0.13.1
matplotlib==3.8.2
xlsxwriter==3.1.9
python-dotenv==1.0.0